Add a weekly cluster config (#600)
Automatic merge from submit-queue.

Add a weekly cluster config

Currently proxy and mixer are auto-updated by using the 'weekly-staging' tag. 

Keel listens to pubsub events generated when a tag is applied to gcr.io and updates the pods 
with the new SHA. 

The 'semantic version' experiment in release02 is currently not working, looking for a way to fix it.

Did a bit of cleanup and update release02 to 0.2.12 manually.

```release-note
none
```